+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
    <modelVersion>4.0.0</modelVersion>
-   <groupId>com.niccholaspage.Fe</groupId>
+   <groupId>com.niccholaspage</groupId>
    <artifactId>Fe</artifactId>
    <version>0.8.3</version>
    <name>Fe</name>
