+++ src/main/java/com/niccholaspage/Fe/command/CommandType.java
@@ -1,0 +1,7 @@
+package com.niccholaspage.Fe.command;
+
+public enum CommandType {
+    CONSOLE,
+    PLAYER,
+    CONSOLE_WITH_ARGUMENTS
+}
