--- src/main/java/org/melonbrew/fe/command/CommandType.java
@@ -1,7 +1,0 @@
-package org.melonbrew.fe.command;
-
-public enum CommandType {
-    CONSOLE,
-    PLAYER,
-    CONSOLE_WITH_ARGUMENTS
-}
