--- src/main/java/org/melonbrew/fe/command/commands/ReloadCommand.java
@@ -1,27 +1,0 @@
-package org.melonbrew.fe.command.commands;
-
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandSender;
-import org.melonbrew.fe.Fe;
-import org.melonbrew.fe.Phrase;
-import org.melonbrew.fe.command.CommandType;
-import org.melonbrew.fe.command.SubCommand;
-
-public class ReloadCommand extends SubCommand {
-    private final Fe plugin;
-
-    public ReloadCommand(Fe plugin) {
-        super("reload", "fe.reload", "reload", Phrase.COMMAND_RELOAD, CommandType.CONSOLE);
-
-        this.plugin = plugin;
-    }
-
-    @Override
-    public boolean onCommand(CommandSender sender, Command cmd, String commandLabel, String[] args) {
-        plugin.reloadConfig();
-
-        Phrase.CONFIG_RELOADED.sendWithPrefix(sender);
-
-        return true;
-    }
-}
